Refactor: 내 캐릭터 꾸미기 페이지 파일 구조 리팩토링 #SCRUM-82
